### 2018-07-24 #833 @whedon
[PRE REVIEW]: diwanalr: An R data analysis package for diffusing-wave spectroscopy
**Submitting author:** @peterjwatkins (<a href="http://orcid.org/0000-0002-1290-3843">Peter Watkins</a>)
**Repository:** <a href="https://github.com/peterjwatkins/diwanalr" target ="_blank">https://github.com/peterjwatkins/diwanalr</a>
**Version:** v0.1
