### 2018-07-24 #833 @arfon
@peterjwatkins - please make sure the YAML header in your paper matches the formatting of this example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
