### 2018-07-25 #833 @peterjwatkins
The YAML header has been reviewed and the problem rectified - incorrect indent on affiliations
