### 2018-07-26 #833 @peterjwatkins
The article PDF is satisfactory
