### 2018-08-04 #833 @arfon
@peterjwatkins - could you take a look at this list of reviewers https://bit.ly/joss-reviewers and suggest a few potential candidates?
