### 2018-08-06 #833 @arfon
@chlalanne @alexpghayes @semacu - would you be willing to review this submission for JOSS?
