### 2018-08-08 #833 @alexpghayes
I'd be happy to. Just FYI: this will be my first review. I'll also be away from the internet intermittently for the next two weeks.
