### 2018-08-09 #833 @semacu
Thanks for letting me know, I am afraid I am not familiar enough with diffusing-wave spectroscopy to be able to do a good job at reviewing this work.
