### 2018-09-14 #947 @arfon
👋 @alexpghayes - please take a look at the review checklist above and work your way through it. If you have any questions for @peterjwatkins as you go, ask them here or make issues on the repository: https://github.com/peterjwatkins/diwanalr

If anything isn't clear, just give me a shout!
