### 2018-09-20 #947 @alexpghayes
## Summary

This package appears to perform a set of useful calculations, but is not yet ready for public consumption.

### Documentation

- I recommend adding the material in the intro notebook / vignette to the Github README. An increasing number of people see the README before anything else. Most R users don't use Jupyter notebooks, so it doesn't make sense for the most public facing introduction to be in a Jupyter notebook.

- I recommend printing the original data and the intermediate data frames throughout the intro (this will happen automatically if you use a `.Rmd` README, which you can setup with `usethis::use_readme_rmd()`). This will help users understand the input and output of the `diwanalr` functions. Similarly, having plots in the README will make it more compelling.

- Single letter variables names are uninformative and considered bad style in R. I would give variables throughout the documentation meaningful names.

 - `calc_g1()` uses a smoothing spline, but based on my reading of the documentation I had expected it to make an exact calculation.

### Function documentation

Some questions I had reading the documentation for various functions:
- `calc_g1()`: What is `n`? What does it do? Do the columns in `d` need to be named? Do they need to be in a particular order?
- `plot_g1()`: What constitutes an acceptable input tibble?
- `calc_modulus`: What are the units of `radius`?

### Provided examples

I get some warnings running the examples in the vignette:

``` r
library(diwanalr)

data(dws)
d <- calc_g1(dws)
gp <- plot_g1(d) ; gp
e <- calc_msd(d)
ep <- plot_msd(e) ; ep
f <- calc_modulus(e)
fp <- plot_modulus(f) ; fp  # gives a warning for me
#> Warning in self$trans$transform(x): NaNs produced
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Removed 165 rows containing missing values (geom_point).
```

Created on 2018-09-18 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).

### R package development practice

- Variables in examples and code should have informative names, rather than just being a single letter
- Function arguments should also have informative names
- Depending on the tidyverse is bad practice and susecptable to breakage -- https://www.tidyverse.org/articles/2018/06/tidyverse-not-for-packages/

### Tests

- The existing tests check the type of objects returned from functions, but they should also check the correctness of the calculations for some simple cases

- The package should pass R CMD Check with no ERRORS or WARNINGS and ideally no NOTES. To run R CMD Check, install the `devtools` package and then run `devtools::check()`. More details are [available here](http://r-pkgs.had.co.nz/).

- The tests should be automated via some sort of continuous integration service. `usethis::use_travis()` can help set this up.

### Community guidelines

- The are no community/contribution guidelines.
- `usethis::use_tidy_contributing()` can help set this up.

### References

I'm not sure what needs a DOI or how to check that one is present.
