### 2018-09-20 #947 @arfon
> cc @peterjwatkins @arfon -- do I need to tag y'all to make sure you get notifications or is just posting to this thread enough?

We all got tagged at the top by @whedon so @peterjwatkins should be getting these notifications :-)
