### 2018-09-21 #947 @peterjwatkins
@peterjwatkins is getting the notifications :-)

Thanks for the review and feedback.. about to go on leave for 3 weeks, and I'll get back when I return.
