### 2018-10-26 #947 @peterjwatkins
Just to advise that I've been away for a little while, and I'm working with the comments from @alexpghayes
