### 2018-10-26 0450930
Modification to R scripts
