+++ man/calc_freq.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/modulus.R
-\name{calc_freq}
-\alias{calc_freq}
-\title{This function is used to calculate the frequency as inverse of time}
-\usage{
-calc_freq(t)
-}
-\arguments{
-\item{t}{A vector}
-}
-\value{
-A vector
-}
-\description{
-This function is used to calculate the frequency as inverse of time
-}
-\examples{
-calc_freq(time)
-}
