+++ man/calc_msd.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/msd.R
-\name{calc_msd}
-\alias{calc_msd}
-\title{Calculate the mean square displacement}
-\usage{
-calc_msd(d)
-}
-\arguments{
-\item{d}{A tibble consisting of correlation time, observed and scaled g1(t) values}
-}
-\value{
-A tibble consisting of correlation time and related mean square displacement
-}
-\description{
-Calculate the mean square displacement
-}
-\examples{
-e <- calc_msd(d)
-}
