### 2018-10-26 171652d
Revise plot_modulus
