+++ man/calc_modulus.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{form_modulus}
-\alias{form_modulus}
+\name{calc_modulus}
+\alias{calc_modulus}
 \title{This function produces a tibble consisting of frequency, the storage and loss modulus}
 \usage{
-form_modulus(msd_t, temp = NULL, radius = NULL)
+calc_modulus(d, temp = NULL, radius = NULL)
 }
 \arguments{
-\item{msd_t}{A tibble consisting of correlation time and related mean square displacement}
+\item{d}{A tibble consisting of correlation time and related mean square displacement}
 
 \item{temp}{(optional) temperature in Celsius, default = 20}
 
