+++ man/calc_GDprime.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{loss_mod}
-\alias{loss_mod}
+\name{calc_GDprime}
+\alias{calc_GDprime}
 \title{This function is used to calculate the loss modulus}
 \usage{
-loss_mod(visco_mod, slope)
+calc_GDprime(G, slope)
 }
 \arguments{
-\item{visco_mod}{A vector (viscoelastic modulus)}
+\item{G}{A vector (viscoelastic modulus)}
 
 \item{slope}{A vector}
 }
@@ -18,5 +18,5 @@ A vector
 This function is used to calculate the loss modulus
 }
 \examples{
-Loss <- loss_mod(visco_mod, slope)
+Loss <- calc_GDprime(G, slope)
 }
