+++ man/calc_slope.Rd
@@ -1,15 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{form_slope}
-\alias{form_slope}
+\name{calc_slope}
+\alias{calc_slope}
 \title{This function is used internally for calculating the viscoelastic modulus.}
 \usage{
-form_slope(time, msd)
+calc_slope(x, y)
 }
 \arguments{
-\item{time}{A vector}
+\item{x}{A vector}
 
-\item{msd}{A vector (same length as x)}
+\item{y}{A vector (same length as x)}
 }
 \value{
 A vector
@@ -18,5 +18,5 @@ A vector
 This function is used internally for calculating the viscoelastic modulus.
 }
 \examples{
-slope(time, msd)
+calc_slope(time, msd)
 }
