+++ man/calc_Gprime.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{storage_mod}
-\alias{storage_mod}
+\name{calc_Gprime}
+\alias{calc_Gprime}
 \title{This function is used to calculate the storage modulus}
 \usage{
-storage_mod(visco_mod, slope)
+calc_Gprime(G, slope)
 }
 \arguments{
-\item{visco_mod}{A vector (viscoelastic modulus)}
+\item{G}{A vector (viscoelastic modulus)}
 
 \item{slope}{A vector}
 }
@@ -18,5 +18,5 @@ A vector
 This function is used to calculate the storage modulus
 }
 \examples{
-Storage <- storage_mod(visco_mod, slope)
+Storage <- calc_Gprime(G, slope)
 }
