+++ man/calc_freq.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{form_freq}
-\alias{form_freq}
+\name{calc_freq}
+\alias{calc_freq}
 \title{This function is used to calculate the frequency as inverse of time}
 \usage{
-form_freq(time)
+calc_freq(t)
 }
 \arguments{
-\item{time}{A vector}
+\item{t}{A vector}
 }
 \value{
 A vector
@@ -16,5 +16,5 @@ A vector
 This function is used to calculate the frequency as inverse of time
 }
 \examples{
-form_freq(time)
+calc_freq(time)
 }
