+++ man/calc_G.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{viscoelastic_mod}
-\alias{viscoelastic_mod}
+\name{calc_G}
+\alias{calc_G}
 \title{This function is used to calculate the viscoelastic modulus.}
 \usage{
-viscoelastic_mod(temp, radius, msd, slope)
+calc_G(temp, radius, msd, slope)
 }
 \arguments{
 \item{temp}{A number (temperature, in Celsius)}
@@ -22,5 +22,5 @@ A vector
 This function is used to calculate the viscoelastic modulus.
 }
 \examples{
-viscoelastic_mod(temp, radius, msd, slope)
+calc_G(temp, radius, msd, slope)
 }
