### 2018-10-26 9b2eb8b
Update R/modulus.R
