### 2018-10-26 61afd52
Revised viscoelastic modulus function
