+++ man/visco_mod.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{calc_G}
-\alias{calc_G}
+\name{visco_mod}
+\alias{visco_mod}
 \title{This function is used to calculate the viscoelastic modulus.}
 \usage{
-calc_G(temp, radius, msd, slope)
+visco_mod(temp, radius, msd, slope)
 }
 \arguments{
 \item{temp}{A number (temperature, in Celsius)}
@@ -22,5 +22,5 @@ A vector
 This function is used to calculate the viscoelastic modulus.
 }
 \examples{
-calc_G(temp, radius, msd, slope)
+visco_mod(temp, radius, msd, slope)
 }
