### 2018-10-26 de6c8ca
Minor changes to modulus script
