+++ man/form_modulus.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{calc_modulus}
-\alias{calc_modulus}
+\name{form_modulus}
+\alias{form_modulus}
 \title{This function produces a tibble consisting of frequency, the storage and loss modulus}
 \usage{
-calc_modulus(d, temp = NULL, radius = NULL)
+form_modulus(msd_t, temp = NULL, radius = NULL)
 }
 \arguments{
-\item{d}{A tibble consisting of correlation time and related mean square displacement}
+\item{msd_t}{A tibble consisting of correlation time and related mean square displacement}
 
 \item{temp}{(optional) temperature in Celsius, default = 20}
 
@@ -20,5 +20,5 @@ A tibble consisting of frequency with related storage and loss modulii
 This function produces a tibble consisting of frequency, the storage and loss modulus
 }
 \examples{
-e <- calc_modulus(d, 20, 5e-7)
+mods <- form_modulus(d, 20, 5e-7)
 }
