+++ man/storage_mod.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{calc_Gprime}
-\alias{calc_Gprime}
+\name{storage_mod}
+\alias{storage_mod}
 \title{This function is used to calculate the storage modulus}
 \usage{
-calc_Gprime(G, slope)
+storage_mod(visco_mod, slope)
 }
 \arguments{
-\item{G}{A vector (viscoelastic modulus)}
+\item{visco_mod}{A vector (viscoelastic modulus)}
 
 \item{slope}{A vector}
 }
@@ -18,5 +18,5 @@ A vector
 This function is used to calculate the storage modulus
 }
 \examples{
-Storage <- calc_Gprime(G, slope)
+Storage <- visco_mod(G, slope)
 }
