+++ man/loss_mod.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/modulus.R
-\name{calc_GDprime}
-\alias{calc_GDprime}
+\name{loss_mod}
+\alias{loss_mod}
 \title{This function is used to calculate the loss modulus}
 \usage{
-calc_GDprime(G, slope)
+loss_mod(visco_mod, slope)
 }
 \arguments{
-\item{G}{A vector (viscoelastic modulus)}
+\item{visco_mod}{A vector (viscoelastic modulus)}
 
 \item{slope}{A vector}
 }
@@ -18,5 +18,5 @@ A vector
 This function is used to calculate the loss modulus
 }
 \examples{
-Loss <- calc_GDprime(G, slope)
+Loss <- loss_mod(visco_mod, slope)
 }
