+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-export(calc_modulus)
 export(form_g1)
+export(form_modulus)
 export(form_msd)
 export(plot_g1)
 export(plot_modulus)
