### 2018-10-26 b0e1c11
Minor changes to modulus plot
