+++ R/modulus.R
@@ -91,12 +91,16 @@ form_modulus <- function(msd_t, temp = NULL, radius = NULL) {
 }
 #' Plots storage and loss modulus against the measured frequency
 #' @param mod_t A tibble consisting of frequency and the related storage and loss modulusA
+#' @param x_max (optional) Maximum 'x' scale value, default = 10000
+#' @param y_max (optional) Maximum 'y' scale value, default = 10000
 #' @examples
 #' plot_modulus(mod_t)
 #' @export
 #' @importFrom tidyr gather
 #' @importFrom ggplot2 ggplot aes geom_point scale_x_log10 scale_y_log10 labs
-plot_modulus <- function(mod_t) {
+plot_modulus <- function(mod_t, x_max = NULL, y_max = NULL) {
+  x_max <- ifelse(is.null(x_max), 10000, x_max)
+  y_max <- ifelse(is.null(y_max), 300, y_max)  
   mod_t <- tidyr::gather(mod_t, key = Modulus, val, -freq)
   mod_p <- ggplot2::ggplot(mod_t, ggplot2::aes(freq, val, color = Modulus)) + 
       ggplot2::geom_point() + 
