### 2018-10-26 364978d
Advice to contributors
