### 2018-10-29 bb51b7b
Moderate changes for paper
