+++ tests/testthat/test_g1.R
@@ -2,12 +2,7 @@ context("testing g1")
 
 library("testthat")
 
-test_that("calc_g1 returns df", {
-  expect_is(calc_g1(dws),  "data.frame") 
+test_that("form_g1 returns df", {
+  expect_is(form_g1(dws),  "data.frame") 
 }
 )
-
-test_that("plot_g1 returns ggplot object",{
-  p <- plot_g1(calc_g1(dws))
-  expect_is(p,"ggplot")
-})
