+++ tests/testthat.R
@@ -1,4 +1,5 @@
-library(testthat)
+library("testthat")
 library(diwanalr)
 
 test_check("diwanalr")
+
