+++ DESCRIPTION
@@ -1,12 +1,14 @@
 Package: diwanalr
 Type: Package
-Title: A set of tools for data from diffusive wave spectroscopy.
+Title: A set of tools for data from diffusive wave spectroscopy
 Version: 0.1.0
 Author: Peter Watkins
 Maintainer: Peter Watkins <Peter.Watkins@csiro.au>
 Imports:
     ggplot2,
-    tidyverse
+    dplyr,
+    tibble, 
+    tidyr
 Description: A set of tools for data resulting from diffusive wave spectroscopy.
     These included conversion of measured data (time vs g2) to related g1, calculation of
     the mean square displacement as well as the storage and long modulus along with related visualisation.

[…]
