### 2018-10-29 #947 @peterjwatkins
Thanks @alexpghayes for the review/feedback in relation to the package.

Some preliminary comments in reply to the review/feedback.

In relation to the documentation, I have made the changes in relation to the documentation; I've added most of the vignette material to the README which, as a result, now has the plots embedded. It does look a lot more compelling. The comment in relation to single letter variables is noted; these have been removed and replaced by more meaningful names, which will be more understandable to the reader. This applies to the function names, along with the arguments to same.

I have clarified the use of the spline approach in the calculation as well.

In relation to the examples, the warnings appear to the presence of negative numbers within the data, and the use of log transformation on the graph scale generates these messages. I have used 'suppressWarnings' in the function to remove these.

In the development practice, I have removed the generic call to the 'tidyverse' and have made explicit calls to the related package needed for the respective functions.

With respect to the tests, the correctness of the calculations needs the use of a standarised set of solutions to use with the DWS technique. This is a work presently in progress.

As suggested, I've been using 'devtools::check()' and there are two issues which are arising; one relates to
checking examples ...
*ERROR
  Running examples in 'diwanalr-Ex.R' failed
  The error most likely occurred in:

  > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
  > ### Name: calc_freq
  > ### Title: This function is used to calculate the frequency as inverse of
  > ###   time
  > ### Aliases: calc_freq
  >
  > ### ** Examples
  >
  > calc_freq(time)
  Error in calc_freq(time) : could not find function "calc_freq"*

which *is* present in the code,

and
*library("testthat")
Error in library("testthat") : there is no package called 'testthat'

  Execution halted*

even though 'testthat' is present.

I'll address the remaining comments subsequently.
