### 2018-10-30 #947 @peterjwatkins
@arfon I have responded to the comments made by @alexpghayes
