### 2018-10-30 0387c0f
Minor updates to g1.R
