+++ README.md
@@ -18,7 +18,7 @@ If needed, the *diwanalr* package can be downloaded using:
 
 One proposed workflow involves the calculation of the storage and loss moduli. The workflow consists of three steps:
 
-*The temporal autocorrelation function*, *g*<sub>1</sub>(*t*). In some cases, the output generated from research equipment consists of a CSV file, with two columns; the first being the correlation time while the second is the measured value of the *intensity autocorrelation function*, noted as *g*<sub>2</sub>(*t*). The *g*<sub>2</sub>(*t*) values are related to *g*<sub>1</sub>(*t*) by the Seigert relationship where *g*<sub>2</sub>(*t*) = 1 + |*g*<sub>1</sub>(*t*)|**<sup>2</sup>. The function, *form\_g1*, accepts a tibble consisting measured correlation times, *t*, and associated *g*<sub>2</sub>(*t*) values, and returns a tibble consisting of three columns; namely, the correlation time (*t*), related *g*<sub>1</sub>(*t*) values as well a set of scaled *g*<sub>1</sub>(*t*) values, ranging from 0 to approximately 1. The scaled values are deployed in the second step of the workflow. Some functionality is provided to allow the user to set an appropriate normalisation value. The function *plot\_g1* is available to visualise the data.
+*The temporal autocorrelation function*, *g*<sub>1</sub>(*t*). In some cases, the output generated from research equipment consists of a CSV file, with two columns; the first being the correlation time while the second is the measured value of the *intensity autocorrelation function*, noted as *g*<sub>2</sub>(*t*). The *g*<sub>2</sub>(*t*) values are related to *g*<sub>1</sub>(*t*) by the Seigert relationship where *g*<sub>2</sub>(*t*) = 1 + |*g*<sub>1</sub>(*t*)|**<sup>2</sup>. The function, *form\_g1*, accepts a tibble consisting measured correlation times, *t*, and associated *g*<sub>2</sub>(*t*) values, and returns a tibble consisting of three columns; namely, the correlation time (*t*), related *g*<sub>1</sub>(*t*) values as well a set of scaled *g*<sub>1</sub>(*t*) values, ranging from 0 to approximately 1. The scaled values are deployed in the second step of the workflow. Some functionality is provided to allow the user to set an appropriate normalisation value. The function *plot\_g1* is available to visualise the data. The *plot\_g1* function uses a spline to estimate t<sub>1/2</sub> for the data.
 
 ``` r
 library(diwanalr)
