+++ README.Rmd
@@ -31,7 +31,7 @@ devtools::install_github("peterjwatkins/diwanalr", force=TRUE)
 One proposed workflow involves the calculation of the storage and loss moduli. The workflow consists of three steps:
 
 *The temporal autocorrelation function*, *g*~1~($t$).
-In some cases, the output generated from research equipment consists of a CSV file, with two columns; the first being the correlation time while the second is the measured value of the *intensity autocorrelation function*, noted as *g*~2~($t$). The *g*~2~($t$) values are related to *g*~1~($t$) by the Seigert relationship where *g*~2~($t$) = 1 + $|$*g*~1~($t$)$|$$^2$. The function, *form_g1*, accepts a tibble consisting measured correlation times, $t$, and associated *g*~2~($t$) values, and returns a tibble consisting of three columns; namely, the correlation time ($t$), related *g*~1~($t$) values as well a set of scaled *g*~1~($t$) values, ranging from 0 to approximately 1. The scaled values are deployed in the second step of the workflow. Some functionality is provided to allow the user to set an appropriate normalisation value. The function *plot_g1* is available to visualise the data.
+In some cases, the output generated from research equipment consists of a CSV file, with two columns; the first being the correlation time while the second is the measured value of the *intensity autocorrelation function*, noted as *g*~2~($t$). The *g*~2~($t$) values are related to *g*~1~($t$) by the Seigert relationship where *g*~2~($t$) = 1 + $|$*g*~1~($t$)$|$$^2$. The function, *form_g1*, accepts a tibble consisting measured correlation times, $t$, and associated *g*~2~($t$) values, and returns a tibble consisting of three columns; namely, the correlation time ($t$), related *g*~1~($t$) values as well a set of scaled *g*~1~($t$) values, ranging from 0 to approximately 1. The scaled values are deployed in the second step of the workflow. Some functionality is provided to allow the user to set an appropriate normalisation value. The function *plot_g1* is available to visualise the data. The *plot_g1* function uses a spline to estimate *t*~1/2~ of the data.
 
 
 ```{r first}
