### 2018-10-31 4a278d6
Some cleanup for R CMD Check
