### 2018-10-31 710c8a2
Everything but the error for examples
