### 2018-10-31 #947 @alexpghayes
I've made some additional comments in peterjwatkins/diwanalr#1, where I did a bit of code cleanup for R CMD Check.
