### 2018-11-04 #947 @arfon
@peterjwatkins - I think this might be ready for you to take another look.
