### 2018-11-07 19c3429
Changes for 'calc_freq'
