### 2019-01-29 #947 @peterjwatkins
@alexpghayes<https://github.com/alexpghayes>

I have had the chance to revisit this, and thee’s an issue with travis with running the examples.

[cid:image002.jpg@01D4B7DF.AAD0AEE0]

I’m not certain where to find ‘diwanalr-Ex.R’ file to sort this error (first time at doing this).. are there any suggestions at all?

Thanks
