### 2019-01-30 #947 @alexpghayes
This means the code in one of the examples isn't self-contained and/or breaks somehow. I recommend using `devtools::run_examples()` to figure out which example is broken.
