+++ tests/testthat.R
@@ -1,4 +1,5 @@
-library("testthat")
+#library("testthat")
+require(testthat)
 library(diwanalr)
 
 test_check("diwanalr")
