+++ man/calc_freq.Rd
@@ -2,19 +2,19 @@
 % Please edit documentation in R/modulus.R
 \name{calc_freq}
 \alias{calc_freq}
-\title{This function is used to calculate the frequency as inverse of time (t)}
+\title{This function is used to calculate the frequency as inverse of time (time_c)}
 \usage{
-calc_freq(t)
+calc_freq(time_c)
 }
 \arguments{
-\item{t}{A vector}
+\item{time_c}{A vector}
 }
 \value{
 A vector
 }
 \description{
-This function is used to calculate the frequency as inverse of time (t)
+This function is used to calculate the frequency as inverse of time (time_c)
 }
 \examples{
-calc_freq(t)
+calc_freq(time_c)
 }
