+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: A set of tools for data resulting from diffusive wave spectroscopy.
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
 Suggests: knitr,
     rmarkdown
 VignetteBuilder: knitr
