+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(calc_freq)
 export(form_g1)
 export(form_modulus)
 export(form_msd)
