### 2019-02-04 bc03fcb
Removed some examples
