### 2019-02-04 8d1bfbe
Edits for R CMD check OK
