### 2019-02-05 392210e
Edits to reduce warning messages
