### 2019-02-05 d60027d
Removes warning note for R CMD check
