+++ R/globals.R
@@ -1,0 +1,6 @@
+# Used to remove warnings for 
+# "no visible binding for global variable"
+#
+#' @import utils 
+utils::globalVariables(c('freq','g1','g2','msd','Modulus','Observed',
+                         'Scaled','time', 'val','Value'))
