### 2019-02-05 #947 @peterjwatkins
@alexpghayes - I've cleaned up the code and R CMD runs OK now, with only 1 note now.. can you review when convenient.
