### 2019-02-07 #947 @alexpghayes
Can you comment on:

https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/R/msd.R#L51

and

https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/R/modulus.R#L110

Why are `NA` values being generated and why is it safe to ignore them?

At this point the code is decently clean but hitting it with [`styler`](https://github.com/r-lib/styler) might make it slightly more readable.

The tests currently only check that calculations are in a `data.frame()`. They should really verify that the calculated values matched with some expected values for easy tests cases. So far it looks like the package was developed on a single dataset, so trying the functions on other datasets would be a good idea (as would using these additional datasets for additional tests).
