### 2019-02-08 10a3905
Comment on use of 'suppressWarnings'
