+++ R/modulus.R
@@ -108,4 +108,7 @@ plot_modulus <- function(mod_t, x_max = NULL, y_max = NULL) {
       ggplot2::scale_y_log10(limits = c(1, 300)) + 
       ggplot2::labs(x = "Frequency", y = "Modulus")
   suppressWarnings(print(mod_p))
+  ## Negative numbers are generated, which are artefacts from the calculation
+  ## which generates warnings when this function is deployed, and is not serious.
+  ## The last statement is used to filter out these statements.
 }
