### 2019-02-08 69f05d6
Used 'styler' to prettify R code
