+++ R/modulus.R
@@ -3,7 +3,7 @@
 #' @param temp A numnber
 #' @return A number
 celsius_to_kelvin <- function(temp) {
-    return(temp + 273.15)
+  return(temp + 273.15)
 }
 #' This function is used internally for calculating the viscoelastic modulus.
 #' @param x A vector
@@ -13,17 +13,17 @@ celsius_to_kelvin <- function(temp) {
 #' @dontrun{
 #' calc_slope(time, msd)
 #' }
-#' 
+#'
 calc_slope <- function(x, y) {
-    ## x and y must be of same length
-    slope <- NULL
-    slope[1] <- NA
-    for (i in 2:(length(x) - 1)) {
-        j <- seq(i - 1, i + 1, 1)
-        slope <- cbind(slope, stats::lm(log(y[j]) ~ log(x[j]))$coef[2])
-    }
-    slope[i + 1] <- NA
-    return(slope)
+  ## x and y must be of same length
+  slope <- NULL
+  slope[1] <- NA
+  for (i in 2:(length(x) - 1)) {
+    j <- seq(i - 1, i + 1, 1)
+    slope <- cbind(slope, stats::lm(log(y[j]) ~ log(x[j]))$coef[2])
+  }
+  slope[i + 1] <- NA
+  return(slope)
 }
 #' This function is used to calculate the viscoelastic modulus.
 #' @param temp A number (temperature, in Celsius)
@@ -32,11 +32,11 @@ calc_slope <- function(x, y) {
 #' @param slope A vector
 #' @return A vector
 visco_mod <- function(temp, radius, msd, slope) {
-    slope <- stats::na.omit(slope)
-    kBoltzmann <- 1.38064852/1e+23
-    lo <- 2
-    high <- length(msd) - 1
-    return((kBoltzmann * celsius_to_kelvin(temp))/(pi * radius * msd[(lo:high)] * (gamma(1 + slope))))
+  slope <- stats::na.omit(slope)
+  kBoltzmann <- 1.38064852 / 1e+23
+  lo <- 2
+  high <- length(msd) - 1
+  return((kBoltzmann * celsius_to_kelvin(temp)) / (pi * radius * msd[(lo:high)] * (gamma(1 + slope))))
 }
 #' This function is used to calculate the frequency as inverse of time (time_to_invert)
 #' @param time_to_invert A vector
@@ -45,10 +45,10 @@ visco_mod <- function(temp, radius, msd, slope) {
 #' calc_freq(time_to_invert)
 #' }
 calc_freq <- function(time_to_invert) {
-    # time_to_invert[1] and time_to_invert[length] are removed
-    lo <- 2
-    high <- length(time_to_invert) - 1
-    return(1/time_to_invert[(lo:high)])
+  # time_to_invert[1] and time_to_invert[length] are removed
+  lo <- 2
+  high <- length(time_to_invert) - 1
+  return(1 / time_to_invert[(lo:high)])
 }
 #' This function is used to calculate the storage modulus
 #' @param visco_mod A vector (viscoelastic modulus)
@@ -59,8 +59,8 @@ calc_freq <- function(time_to_invert) {
 #' Storage <- visco_mod(G, slope)
 #' }
 storage_mod <- function(visco_mod, slope) {
-    slope <- stats::na.omit(slope)
-    return(visco_mod * cos(pi * slope/2))
+  slope <- stats::na.omit(slope)
+  return(visco_mod * cos(pi * slope / 2))
 }
 #' This function is used to calculate the loss modulus
 #' @param visco_mod A vector (viscoelastic modulus)
@@ -71,8 +71,8 @@ storage_mod <- function(visco_mod, slope) {
 #' Loss <- loss_mod(visco_mod, slope)
 #' }
[…]
