+++ R/msd.R
@@ -1,5 +1,5 @@
 #--------------
-# Functions for the calculation of the MSD for transmission 
+# Functions for the calculation of the MSD for transmission
 # Credit and acknowledgement to W.N. (Bill) Venables who
 # provided this code/solution
 
@@ -9,25 +9,27 @@
 #' @param y A number
 #' @return A function localised in scope
 msd_g1_diff <- local({
-    lambda <- 6.32/1e+07
-    L <- 0.01
-    lstar <- z0 <- 0.000661
-    k0 <- 2 * pi/lambda
-    function(x, y) {
-        ((((((L/lstar) + 4/3)/((z0/lstar) + 2/3)) * (sinh((z0/lstar) * x) + (2/3) * x * cosh((z0/lstar) * x)))/((1 +
-            (4/9) * x^2) * sinh((L/lstar) * x) + (4/3) * x * cosh((L/lstar) * x))) - y)
-    }
+  lambda <- 6.32 / 1e+07
+  L <- 0.01
+  lstar <- z0 <- 0.000661
+  k0 <- 2 * pi / lambda
+  function(x, y) {
+    ((((((L / lstar) + 4 / 3) / ((z0 / lstar) + 2 / 3)) * (sinh((z0 / lstar) * x) + (2 / 3) * x * cosh((z0 / lstar) * x))) / ((1 +
+      (4 / 9) * x^2) * sinh((L / lstar) * x) + (4 / 3) * x * cosh((L / lstar) * x))) - y)
+  }
 })
 #' Used internally for calculating mean square displacement for transmission geometry
 #' @param y used internally to find root
 #' @return The root value
 #' @importFrom stats uniroot
 findX <- function(y) {
-    tst <- msd_g1_diff(c(.Machine$double.eps, 1), y)
-    if (prod(tst) < 0) {
-        ## opposite signs
-        stats::uniroot(msd_g1_diff, c(.Machine$double.eps, 1), y = y)$root
-    } else NA
+  tst <- msd_g1_diff(c(.Machine$double.eps, 1), y)
+  if (prod(tst) < 0) {
+    ## opposite signs
+    stats::uniroot(msd_g1_diff, c(.Machine$double.eps, 1), y = y)$root
+  } else {
+    NA
+  }
 }
 #' Used internally for calculating mean square displacement for transmission geometry
 #' @param y used internally to find root
@@ -44,25 +46,27 @@ FindX <- Vectorize(findX)
 #' @importFrom stats na.omit
 #' @export
 form_msd <- function(t_g1) {
-    lambda <- 6.32/1e+07
-    k0 <- 2 * pi/lambda
-    g1_msd <- within(t_g1, msd <- FindX(Scaled)/(10^8 * k0))
-    g1_msd <- dplyr::select(g1_msd, -`Observed`, -`Scaled`)
-    g1_msd <- stats::na.omit(g1_msd)
-    ## Note : NA are introduced in the 'findX' function (see above)
-    ## the last statement is used as a filter for the NAs's
-    return(g1_msd)
+  lambda <- 6.32 / 1e+07
+  k0 <- 2 * pi / lambda
+  g1_msd <- within(t_g1, msd <- FindX(Scaled) / (10^8 * k0))
+  g1_msd <- dplyr::select(g1_msd, -`Observed`, -`Scaled`)
+  g1_msd <- stats::na.omit(g1_msd)
+  ## Note : NA are introduced in the 'findX' function (see above)
+  ## the last statement is used as a filter for the NAs's
+  return(g1_msd)
 }
 #' Plots the mean square displacement against the correlation time
 #' @param g1_msd A tibble consisting of correlation time and related mean square displacement
[…]
