+++ R/globals.R
@@ -1,6 +1,8 @@
-# Used to remove warnings for 
+# Used to remove warnings for
 # "no visible binding for global variable"
 #
-#' @import utils 
-utils::globalVariables(c('freq','g1','g2','msd','Modulus','Observed',
-                         'Scaled','time', 'val','Value'))
+#' @import utils
+utils::globalVariables(c(
+  "freq", "g1", "g2", "msd", "Modulus", "Observed",
+  "Scaled", "time", "val", "Value"
+))

[…]
