### 2019-02-08 #947 @peterjwatkins
Thanks @alexpghayes for the feedback.

I've added comments to the files (to make it explicit for the code readers) about these lines -
re: msd.R#L51, I've added as explanations:
Note : NAs are introduced in the 'findX' function and the last statement is used as a filter for the NAs's
while for modulus.R#L110, I've added:
Negative numbers are generated, which are artefacts from the calculation which generates warnings when this function is deployed, and is not serious. The last statement is used to filter out these statements.

The main R code has also been prettified using the 'styler' package as well.

The last point is noted, and is accurate. There is only one dataset  present in the package, used to develop the package. Other datasets are being/will be generated, that will be included with further/future package development. This is a work in progress. Thanks for the feedback.
