### 2019-02-09 #947 @alexpghayes
Quick note on comments:

> Note : NAs are introduced in the 'findX' function and the last statement is used as a filter for the NAs's

This is useful in that it describes *what* is happening, but it doesn't describe *why* it is happening. If someone read this code and was trying to decide if this where a reasonable decision, they'd have to go back and do a bunch of math themselves. I recommend describing exactly which inputs get turned into `NA`s and why it is safe to ignore that input. Without understanding the methods being used here, this feels a lot like throwing out inconvenient data, which makes me suspicious. Some more detailed comments about this could help users trust the software more. Same goes for the second comment.
