### 2019-02-11 169afe7
Added note on NA used in uniroot for findX
