+++ R/msd.R
@@ -24,6 +24,9 @@ msd_g1_diff <- local({
 #' @importFrom stats uniroot
 findX <- function(y) {
   tst <- msd_g1_diff(c(.Machine$double.eps, 1), y)
+  ## For 'uniroot' to behave properly, it's important that f(lower) * f(upper) < 0
+  ## This test ensures that a solution is present to find X
+  ## otherwise, there is no root and thus NA
   if (prod(tst) < 0) {
     ## opposite signs
     stats::uniroot(msd_g1_diff, c(.Machine$double.eps, 1), y = y)$root
@@ -52,7 +55,8 @@ form_msd <- function(t_g1) {
   g1_msd <- dplyr::select(g1_msd, -`Observed`, -`Scaled`)
   g1_msd <- stats::na.omit(g1_msd)
   ## Note : NA are introduced in the 'findX' function (see above)
-  ## the last statement is used as a filter for the NAs's
+  ## meaning that there is no solution to 'uniroot' and thus NA
+  ## The above statement is used as a filter for the NAs's
   return(g1_msd)
 }
 #' Plots the mean square displacement against the correlation time
