### 2019-02-12 9251047
Comment on plot_modulus and findX
