### 2019-02-12 #947 @peterjwatkins
OK, got it now :-)

In relation to the first comment, the introduction of NA relates to 'uniroot'; NA is introduced when there is no solution for 'uniroot' - I've added the following
  ## Otherwise, there is no root and thus NA
which explains why NA is introduced, and then subsequently filtered.

In relation to the second comment; I've revised the comments to:
  ## These are artefacts, and have no impact on the visualisation in the plot
This function is intended as a tool for the end package user to visualise the calculated G' and G''. Inherently, there are negative numbers computed for these values which causes the warnings to appear, without suppression. From a coding viewpoint, I acknowledge the point but the users end, this tool is used for data visualisation, and later use.
