### 2019-02-13 30c1ffc
Modulus values <0.01 are filtered for plots
