### 2019-02-13 19c27c4
Added 'y_threshold' as parameter for modulus plot
