### 2019-03-16 #947 @alexpghayes
Much improved! For the comment

```
 For 'uniroot' to behave properly, it's important that f(lower) * f(upper) < 0.
 This test ensures that a solution is present to find X. Otherwise, there is no root and thus NA
```

it's worth adding whether or not we expect a solution to exist, and why. In both cases, it's worth interpreting what this means in terms of the scientific data in addition to the math.

At https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/R/modulus.R#L110 we have

```
  ## NaNs are generated when plotting modulus due to negative numbers
  ## in G' and G'', which causes warnings with ggplot.
  ## These are artefacts, and have no impact on the visualisation in the plot
```

It looks like modulus values are being plotted directly, and then when the modulus is negative the `scale_*_log()` is breaking because we try to take the log of a negative value. If ignoring negative modulus is fine, it's worth adding a note on this, because otherwise this is just throwing away data. Do we expect modulus to always be nonnegative and then numerical error sometimes makes it negative?
