### 2019-04-24 0405fe6
Minor edit for comment in modulus.R
