### 2019-05-13 #947 @peterjwatkins
@labarba - here's the update :)

For the plot_modulus function, I have revised the preliminary comments to
  ## only positive values are of interest; no data loss occurs with this function.
which, I believe, satisfies the second remark from Alex.

In relation to the first comment, I'm not able to respond to this one as requested. The DWS technique is not my area of expertise; my involvement in this area was developing the software so I can't respond whether a solution exists or otherwise in these cases. My guess is that this would also dependent on the actual systems are being studied using the DWS technique, and will become more evident with further datasets. Consequently, I'm not able to provide any further revision in this case.

Alex also notes that there's an issue relating to tests of completeness, needing simpler datasets. This is still a work in progress with this data to be collected for testing. Thus, I acknowledge that there is a deficit in this case but this will be revised and updated with future revisions of the package.
