### 2019-05-27 #947 @alexpghayes
@peterjwatkins If you don't have real world data sets to use in tests, you can simulate fake data and use the fake data to confirm that calculations are correct.

@labarba Do I make an accept / reject recommendation at some point or just evaluate if the review checklist items have been completed?
