### 2019-05-30 3dd01e2
Added tests for g1 calculations
