### 2019-05-30 c58a80d
msd test calculations added
