### 2019-05-31 1bb8fb4
Re-format of code, and added tests for 'modulus'
