+++ R/g1.R
@@ -27,8 +27,9 @@ form_g1 <- function(t_g2, num_points = NULL) {
   num_points <- ifelse(is.null(num_points), 30, num_points)
   t_g2 <- dplyr::filter(t_g2, g2 >= 1)
   t_g1 <- dplyr::mutate(t_g2, Observed = sqrt(g2 - 1))
-  t_g1 <- dplyr::select(t_g1, -g2)
-  t_g1 <- dplyr::mutate(t_g1, Scaled = range_scale(Observed, num_points)) # range scale data
+  t_g1 <- dplyr::select(t_g1,-g2)
+  t_g1 <-
+    dplyr::mutate(t_g1, Scaled = range_scale(Observed, num_points)) # range scale data
   return(t_g1)
 }
 #' Plots the scaled and observed g1(t)) against the correlation time, highlighting the point where g1(t) = 0.5
@@ -40,15 +41,20 @@ form_g1 <- function(t_g2, num_points = NULL) {
 #' @importFrom ggplot2 ggplot aes geom_point scale_x_log10 labs geom_hline geom_vline
 plot_g1 <- function(t_g1) {
   pred_spl <- g1_spline(t_g1)
-  t_g1_data <- tidyr::gather(t_g1, key = g1, Value, -time)
-  t_g1_plot <- ggplot2::ggplot(t_g1_data, ggplot2::aes(time, Value, color = g1)) +
+  t_g1_data <- tidyr::gather(t_g1, key = g1, Value,-time)
+  t_g1_plot <-
+    ggplot2::ggplot(t_g1_data, ggplot2::aes(time, Value, color = g1)) +
     ggplot2::geom_point() +
     ggplot2::scale_x_log10() +
     ggplot2::labs(x = "Correlation time (s)", y = "g1(t)") +
-    ggplot2::geom_hline(yintercept = 0.5, linetype = "dashed", color = "blue") +
-    ggplot2::geom_vline(xintercept = stats::approxfun(
-      x = pred_spl,
-      y = t_g1$time
-    )(0.5), linetype = "dashed", color = "blue")
+    ggplot2::geom_hline(yintercept = 0.5,
+                        linetype = "dashed",
+                        color = "blue") +
+    ggplot2::geom_vline(
+      xintercept = stats::approxfun(x = pred_spl,
+                                    y = t_g1$time)(0.5),
+      linetype = "dashed",
+      color = "blue"
+    )
   print(t_g1_plot)
 }
