### 2019-06-10 #947 @alexpghayes
Notes on the tests:

- https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/tests/testthat/test_g1.R#L10 does not check any code in the package itself. In general tests should call package functions.
- The tolerance for the tests is very high (0.1 or 0.01 instead of the typically machine precision of around 1e-16. Even loose tests in the numeric world still use tolerances around 1e-6 to 1e-8). Can you comment on why this is the case?
- https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/tests/testthat/test_g1.R#L10 checks only three rows out of a much larger data frame. It would be better to test against all the values in the data frame. Similarly at https://github.com/peterjwatkins/diwanalr/blob/alexpghayes-r_cmd_cleanup_master/tests/testthat/test_msd.R#L27.

Once these are addressed I'm happy to recommend publication.
