### 2019-06-11 00dc99b
New tests for test_g1.R
