### 2019-06-11 #947 @peterjwatkins
In relation to the notes:

- The tests for test_g1.R have been revised. The tests are performed on data generated on the model data, rather than specific calculations as was done previously.

- As noted, the tolerance is high and it is also been revised to 0.05. There is considerable variation in the  result (1.00 +/- 0.05) which is calculated using normalisation of the original data. Using typical machine precision would cause the test to fail, even with precision values of 1e-6 to 1e-8. My aim was to demonstrate the calculations provide a result to close to 1.00 for the early part of the data and close to 0.00 for the latter part of the data. I am satisfied that these tests are suitable and fit-for-purpose. In the event where higher precision is needed for testing the calculations, then I acknowledge and agree that more stricter tolerances would be needed (as demonstrated in test_msd.R, for example).
- The comment is noted but I believe that these tests are sufficient. The tests are meant to identify whether the calculation is providing the correct result. The tests are passing which indicates that the result lies within the required precision for these values, and thus are fit-for-purpose and also sufficient. In this case, I feel that by adding more would not add further value to the purpose of testing of the package.
