### 2019-06-11 951fe8b
Edits to global.R
