### 2019-06-11 e91396d
Minor edit for test_modulus.R
