### 2019-06-11 #947 @alexpghayes
Yep! Happy to recommend for publication!
