### 2019-06-12 #947 @arfon
Hi @peterjwatkins - we're nearly ready to accept here. Before we do so, could you make a few updates to your paper:

- If possible, I think it might read better if the large equation was broken out by itself.
- Would it be possible to expand a little on the introduction/statement of need section? We ask that authors include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader.
- Are there examples of existing academic work that make use of this package? If so, could you cite them here?
