### 2019-06-13 #947 @peterjwatkins
Hi @arfon, as requested, I've made updates to the paper:

- the large equation has been separated and also centered
- I've added some further detail to the introduction as follows:
_____________________________
DWS is an optical technique derived from dynamic light scattering which studies the dynamics of scattered light [Weitz].

If carefully calibrated, DWS allows the quantitative measurement of microscopic motion in a soft material, which can be used with micro-rheology to determine the rheological properties of such a complex medium.

DWS has been applied the analysis to food systems. Some examples include understanding the dynamics of emulsion  systems (e.g. fat in water) such as mayonnaise [Kim 2019] along with the gellation of food products [Alexander 2006].

DWS also has practical applications in the pharmaceutical and cosmetic industries as well [@Reufer2014].

In some instances, the analysis of data resulting from DWS measurements can be tedious depending on, for example, the use of spreadsheets for numerical analysis, or multiple software packages [Niederquell 2017]. The **diwanalr** package has been developed as an alternative approach, allowing data analysis to be readily performed using a single software package, within R. The package has a set of functions which can be readily and seamlessly deployed after the measurement of data. Further functionality can be added to the package as required.
____________________________________

- Unfortunately, there is no publically available work at this stage which uses the package, so I am unable to include any citations at this stage.

I trust that this is satisfactory.
