### 2019-06-13 0b268f9
Updates to paper and bibliography
