+++ paper/paper.bib
@@ -15,3 +15,61 @@
     title = {{Diffusing-wave Spectroscopy}},
     year = 1993
 }
+
+
+@article {Kim2019,
+	author = {Kim, Ha Seong and {\c S}enbil, Nesrin and Zhang, Chi and Scheffold, Frank and Mason, Thomas G.},
+	title = {Diffusing wave microrheology of highly scattering concentrated monodisperse emulsions},
+	volume = {116},
+	number = {16},
+	pages = {7766--7771},
+	year = {2019},
+	doi = {10.1073/pnas.1817029116},
+	publisher = {National Academy of Sciences},
+	issn = {0027-8424},
+	URL = {https://www.pnas.org/content/116/16/7766},
+	eprint = {https://www.pnas.org/content/116/16/7766.full.pdf},
+	journal = {Proceedings of the National Academy of Sciences}
+}
+
+@article{Alexander2006,
+title = "Diffusing wave spectroscopy of gelling food systems: The importance of the photon transport mean free path (l*) parameter",
+journal = "Food Hydrocolloids",
+volume = "20",
+number = "2",
+pages = "325 - 331",
+year = "2006",
+note = "7th International Hydrocolloids Conference",
+issn = "0268-005X",
+doi = "https://doi.org/10.1016/j.foodhyd.2005.02.021",
+url = "http://www.sciencedirect.com/science/article/pii/S0268005X05001189",
+author = "Marcela Alexander and Milena Corredig and Douglas G. Dalgleish",
+keywords = "Milk, Casein micelles, GDL, Acidification, Whey protein, Pectin, Diffusing wave spectroscopy, *"
+}
+
+@article{Reufer2014,
+title = "Introducing Diffusing Wave Spectroscopy as a Process Analytical Tool for Pharmaceutical Emulsion Manufacturing",
+journal = "Journal of Pharmaceutical Sciences",
+volume = "103",
+number = "12",
+pages = "3902 - 3913",
+year = "2014",
+issn = "0022-3549",
+doi = "https://doi.org/10.1002/jps.24197",
+url = "http://www.sciencedirect.com/science/article/pii/S0022354915302902",
+author = "Mathias Reufer and Alexandra H.E. Machado and Andreas Niederquell and Katharina Bohnenblust and Beat Müller and Andreas Charles Völker and Martin Kuentz",
+keywords = "emulsions, polymer, gel, manufacturing, process analytical technology (PAT), physical characterization, microrheology, light scattering, diffusing wave spectroscopy (DWS)",
+}
+
+@article{Niederquell2017,
+title = "A diffusing wave spectroscopy study of pharmaceutical emulsions for physical stability assessment",
+journal = "International Journal of Pharmaceutics",
+volume = "530",
+number = "1",
+pages = "213 - 223",
+year = "2017",
+issn = "0378-5173",
+doi = "https://doi.org/10.1016/j.ijpharm.2017.07.038",
+url = "http://www.sciencedirect.com/science/article/pii/S0378517317306373",
+author = "Andreas Niederquell and Alexandra H.E. Machado and Martin Kuentz"
+}
