### 2019-06-13 #947 @arfon
Thanks @peterjwatkins. That alignment didn't quite work properly - but this fix makes things good I think: https://github.com/peterjwatkins/diwanalr/pull/2
