### 2019-06-14 #947 @peterjwatkins
Thanks @arfon -  as an FYI, I pulled the fix down into R, and under preview, the equation does not display. However, it would be better to compile it as a PDF and see if it works  it might be just a quirk.
