### 2019-06-14 #947 @arfon
@peterjwatkins ☝️this looks good to me. What about you? Although I failed to add a space before the reference (fixed in https://github.com/peterjwatkins/diwanalr/pull/3)
