### 2019-06-14 #947 @peterjwatkins
@arfon the only minor thing is that the first 'g' is not rendered correctly - instead of _g_ it is *g* - might be better to remove the '*' - only a small detail
