### 2019-06-14 #947 @arfon
> @arfon the only minor thing is that the first 'g' is not rendered correctly - instead of g it is * g * - might be better to remove the '*' - only a small detail

https://github.com/peterjwatkins/diwanalr/pull/3 fixes this too now I think.
