### 2019-06-15 #947 @arfon
@peterjwatkins could you merge https://github.com/peterjwatkins/diwanalr/pull/3 please?
