### 2019-06-17 157d9cf
Minor edit to paper.
