### 2019-06-18 cec035b
Version change 0.5.0
