### 2019-06-18 #947 @peterjwatkins
@arfon - I've released the software and it has been archived on Zenodo,  and the DOI is 10.5281/zenodo.3427971
