### 2019-06-18 #947 @arfon
Hrm, looks like that DOI isn't resolving yet (http://doi.org/10.5281/zenodo.3427971), can you confirm you've hit the 'publish' button on Zenodo?
