### 2019-06-19 #947 @peterjwatkins
@arfon.. it turns out, I hadn't :-) - first time that I have used Zenodo.. I've edited the page and hit the 'publish' button on Zenodo.. not certain on how long it will take for the DOI to resolve.
