### 2019-06-19 #947 @arfon
> @arfon.. it turns out, I hadn't :-) - first time that I have used Zenodo.. I've edited the page and hit the 'publish' button on Zenodo.. not certain on how long it will take for the DOI to resolve.

Thanks! It's still not resolving but I'm sure it will soon. I'll set myself a reminder.
