### 2019-06-24 #947 @peterjwatkins
@arfon - I've realised what the issue is - I made a transcription error - the correct URL is http://doi.org/10.5281/zeonodo.3247971 - this is correctly resolving
